+++ src/scripts/main.js
@@ -1,34 +1,9 @@
-'use strict';
-
-const BASE_URL = 'https://mate-academy.github.io/phone-catalogue-static/api';
-const ENDPOINTS = {
-  phones: '/phones.json',
-  phoneById: id => `/phones/${id}.json`,
-};
-
-function request(url) {
-  return fetch(url)
-    .then(response => Promise.resolve(response.json()));
-}
-
-function getPhones() {
-  return request(`${BASE_URL}${ENDPOINTS.phones}`);
-}
+import { getPhones, getPhonesDetails } from './fetchData';
 
 function getIds(phones) {
-  const result = [];
-
-  for (const phone of phones) {
-    result.push(phone.id);
-  }
-
-  return result;
-}
-
-function getPhonesDetails(ids) {
-  return Promise.all(ids.map(id => {
-    return request(`${BASE_URL}${ENDPOINTS.phoneById(id)}`);
-  }));
+  return phones.reduce((accumulator, current) => {
+    return [...accumulator, current.id];
+  }, []);
 }
 
 function showNames(phones) {
